# api.yaml
openapi: 3.0.2
info:
  title: DBS Programmierpraktikum - Phase 4 - API-Spezifikation
  description: >-
    In der folgenden Liste finden Sie die <b>für die Anwendergruppen
    spezifischen Grundfunktionalitäten</b>. Es müssen <b>alle folgenden
    allgemeinen Bedingungen</b> eingehalten werden:
    <ul>
      <li><b>Bei einer gültigen <code>GET</code>-Anfrage wird ein JSON-Array von JSON-Objekten, wobei ein JSON-Objekt eine Entität darstellt, zurückgegeben</b> und dies mit einem Code <code>200</code> belegt. Wird die <code>GET</code>-Anfrage auf einer nicht vorhandenen Ressource ausgeführt, so wird mit Code <code>404</code> geantwortet. Ist die Ressource woanders zu finden, wird dies mit dem Code <code>301</code> und dem Link im Header belegt. Eine <code>GET</code>-Anfrage auf einer Ressource, die eine Sammlung von Entitäten darstellt, kann durch Übergabe von bestimmten Parametern und dazugehörigen Werten <b>case-insensitive gefiltert</b> werden. Dabei muss jede Entität in der Sammlung allen Parametern gleichzeitig genügen. Die Sammlung wird in der Antwort in ein JSON-Array gekapselt.</li>
...
      <li><b>Bei einer gültigen <code>DELETE</code>-Anfrage wird der Code <code>204</code> zurückgegeben.</b> Ist die Ressource nicht vorhanden, wird dies mit Code <code>404</code> belegt.</li>
...
      <li><b>Antworten auf gültige <code>POST</code>- und <code>PATCH</code>-Anfragen beinhalten nichts.</b></li>
...
      <li><b>Die erwähnten IDs sind RowIDs!!!</b></li>
    </ul>
...
servers:
...
paths:
  /kunden:
    get:
      tags:
        - Anwender
      summary: Gibt Kunden zurück.
      description: >-
        Alle **Kunden** werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: beitragsbefreit
          in: query
          description: 'Nur Kunden, die vom Jahresbeitrag **befreit** sind (`true` = beitragsfreit / `false` nicht beitragsbefreit).'
          schema:
            type: boolean
        - name: guthaben
          in: query
          description: 'Nur Kunden, deren **Guthaben** >= `guthaben` ist, werden zurückgegeben.'
          schema:
            type: number
            format: double
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        kundenid:
                          type: integer
                          description: '**RowID des Kunden**.'
                      required:
                        - kundenid
                    - $ref: '#/components/schemas/kunde'
                    - $ref: '#/components/schemas/user'
    post:
      tags:
        - Anwender
      summary: Fügt einen Kunden hinzu.
      description: >-
        Ein **Kunde** wird mit den entsprechenden Werten hinzugefügt. Ein zugeordneter User wird ebenfalls angelegt.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/user'
                - $ref: '#/components/schemas/kunde'
                - type: object
                  properties:
                    adresseid:
                      type: integer
                      description: '**RowID** der Adresse.'
                  required:
                    - adresseid
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
  /bibliothekare:
    get:
      tags:
        - Anwender
      summary: Gibt Bibliothekare zurück.
      description: >-
        Alle **Bibliothekare** werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: telefonnummer
          in: query
          description: 'Nur Bibliothekare, deren **Telefonnummer** `telefonnummer` enthalten, werden zurückgegeben.'
          schema:
            type: string
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        bibliothekarid:
                          type: integer
                          description: '**RowID des Bibliothekars**.'
                      required:
                        - bibliothekarid
                    - $ref: '#/components/schemas/bibliothekar'
                    - $ref: '#/components/schemas/user'
    post:
      tags:
        - Anwender
      summary: Fügt einen Bibliothekar hinzu.
      description: >-
        Ein **Bibliothekar** wird mit den entsprechenden tWerten hinzugefügt. Ein zugeordneter User wird ebenfalls angelegt.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/user'
                - $ref: '#/components/schemas/bibliothekar'

      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
  /genres:
    post:
      tags:
        - Authentifizierter und als Bibliothekar authorisierter Anwender
      security:
        - basicAuth: []
      summary: Fügt ein Genre hinzu.
      description: >-
        Ein **Genre** wird mit den entsprechenden Werten hinzugefügt.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/genre'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
    get:
      tags:
        - Anwender
      summary: Gibt alle Genres zurück.
      description: >-
        Alle **Genres** werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: bezeichnung
          in: query
          description: 'Nur Genres, deren **Bezeichnung** `bezeichnung` enthalten, werden zurückgegeben.'
          schema:
            type: string
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        genreid:
                          type: integer
                          description: '**RowID des Genres**.'
                      required:
                        - genreid
                    - $ref: '#/components/schemas/genre'
  /artikel:
    post:
      tags:
        - Authentifizierter und als Bibliothekar authorisierter Anwender
      security:
        - basicAuth: []
      summary: Fügt einen Artikel hinzu.
      description: >-
        Ein **Artikel** wird mit den entsprechenden Werten hinzugefügt.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - type: object
                  properties:
                    autorid:
                      type: integer
                      description: '**RowID** des Autors.'
                    genreid:
                      type: integer
                      description: '**RowID** des Genres.'
                    mediumid:
                      type: integer
                      description: '**RowID** des Mediums.'
                  required:
                    - autorid
                    - genreid
                    - mediumid
                - $ref: '#/components/schemas/artikel'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
    get:
      tags:
        - Anwender
      summary: Gibt Artikel zurück.
      description: >-
        Alle **Artikel** werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: isbn
          in: query
          description: 'Nur Artikel, deren **ISBN-Nummer** `isbn` enthalten, werden zurückgegeben.'
          schema:
            type: string
        - name: bezeichnung
          in: query
          description: 'Nur Artikel, deren **Bezeichnung** `bezeichnung` enthalten, werden zurückgegeben.'
          schema:
            type: string
        - name: beschreibung
          in: query
          description: 'Nur Artikel, deren **Beschreibung** `beschreibung` enthalten, werden zurückgegeben.'
          schema:
            type: string
        - name: coverbild
          in: query
          description: 'Nur Artikel, deren **Coverbild** `coverbild` heißen, werden zurückgegeben.'
          schema:
            type: string
        - name: erscheinungsdatum
          in: query
          description: 'Nur Artikel, deren **Erscheinungsdatum** >= `erscheinungsdatum` ist, werden zurückgegeben.'
          schema:
            type: string
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        artikelid:
                          type: integer
                          description: '**RowID des Artikels**.'
                      required:
                        - artikelid
                    - $ref: '#/components/schemas/artikel'
  /exemplare:
    post:
      tags:
        - Authentifizierter und als Bibliothekar authorisierter Anwender
      security:
        - basicAuth: []
      summary: Fügt ein Exemplar hinzu.
      description: >-
        Ein **Exemplar** wird mit den entsprechenden Werten hinzugefügt.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - type: object
                  properties:
                    artikelid:
                      type: integer
                      description: '**RowID** des Artikels.'
                  required:
                    - artikelid
                - $ref: '#/components/schemas/exemplar'
                - $ref: '#/components/schemas/standort'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
    get:
      tags:
        - Anwender
      summary: Gibt Exemplare zurück.
      description: >-
        Alle **Exemplare** werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: preis
          in: query
          description: 'Nur Exemplare, deren **Anschaffungspreis** >= `preis` ist, werden zurückgegeben.'
          schema:
            type: number
            format: double
        - name: ausgeliehen
          in: query
          description: 'Nur Exemplare, die aktuell ausgeliehen (`ausgeliehen` = **true**) bzw. nicht ausgeliehen sind (`ausgeliehen` = **false**) sind, werden zurückgegeben.'
          schema:
            type: boolean
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        exemplarid:
                          type: integer
                          description: '**RowID des Exemplars**.'
                        artikelid:
                          type: integer
                          description: '**RowID des Artikels**.'
                      required:
                        - exemplarid
                        - artikelid
                    - $ref: '#/components/schemas/exemplar'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  responses:
    '201':
      description: '**Created**.'
      headers:
        Location:
          description: Die URL der neuen Ressource.
          schema:
            type: string
...
    '400':
      description: '**Bad Request**.'
...
  schemas:
...
    bibliothekar:
      type: object
      properties:
        telefonnummer:
          type: string
          description: '**Telefonnummer** des Bibliothekars.'
      required:
        - telefonnummer
...
    kunde:
      type: object
      properties:
        guthaben:
          type: number
          format: double
          description: '**Guthaben** des Kunden.'
        beitragsbefreit:
          type: boolean
          description: 'Vom **Jahresbetrag befreit** (befreit = `true` oder nicht befreit = `false`).'
      required:
        - guthaben
        - beitragsbefreit
...